getSampleData {imotionsData}	R Documentation

Downloads a sample as a data.frame

Description

Available samples in your study can be listed using the listSampleSources and listRespondentSamples functions.

Usage

getSampleData(
  connection,
  studyId,
  stimulusId,
  respondentId = NULL,
  segmentId = NULL,
  sampleId,
  signals = NULL
)

Arguments

connection
As returned from connect function.

studyId
The study of interest

stimulusId
The stimulus of interest

respondentId
The respondent to query (only if segmentId = NULL)

segmentId
The segment to query (only if respondentId = NULL)

sampleId
The sample to query

signals
Optional: A vector of signals you would like to return in the data.frame

Details

To get access to sample data directly as a tab-delimited file, use imotionsData::getSampleDataFilePath.

Samples are a portion of the data gathered for a given respondent during exposure to a stimuli, from a given sample source.

Value

A data.frame containing the sample in question.

Examples

studies <- imotionsData::listStudies(connection)
studyId <- studies$id[1]
stimuli <- imotionsData::listStimuli(connection, studyId)
stimulus <- stimuli[1,]
respondents <- imotionsData::listStimuliRespondents(connection, studyId, stimulus$id)
sampleSources <- imotionsData::listSampleSources(connection, studyId, stimulus$id)
sampleData <- imotionsData::getSampleData(connection, studyId, stimulus$id, respondentId = respondents$id[1],
 sampleId = sampleSources$id[1])

[Package imotionsData version 1.8.5 Index]