getSampleDataFilePath {imotionsData}	R Documentation

Downloads a sample as a tab-delimited file.

Description

Consume the file from your tool of choice, for example readr or data.table.

Usage

getSampleDataFilePath(
  connection,
  studyId,
  stimulusId,
  respondentId = NULL,
  segmentId = NULL,
  sampleId
)

Arguments

connection
As returned from connect function.

studyId
The study id

stimulusId
The stimulus for which you would like to load data

respondentId
The respondent to query (only if segmentId = NULL)

segmentId
The segment to query (only if respondentId = NULL)

sampleId
The sample to query

Details

Samples can be listed using the listSampleSources and listRespondentSamples functions.

Samples are a portion of the data collected for a given respondent during exposure to a stimuli, from a given sample source.

Value

The path to the file in question on your local hard-disk.

Examples

studies <- imotionsData::listStudies(connection)
studyId <- studies$id[1]
stimuli <- imotionsData::listStimuli(connection, studyId)
stimulus <- stimuli[1,]
respondents <- imotionsData::listStimulusRespondents(connection, studyId, stimulus$id)
sampleSources <- imotionsData::listSampleSources(connection, studyId, stimulus$id)
sampleFilePath <- imotionsData::getSampleDataFilePath(connection, studyId, stimulus$id, sampleSources$name[1], respondents$id[1])

[Package imotionsData version 1.8.5 Index]