getStudyId {imotionsData}	R Documentation

Retrieves the id of a study given its name

Description

Retrieves the id of a study given its name

Usage

getStudyId(connection, studyName)

Arguments

connection
As returned from connect function.

studyName
The name of a study uploaded to the cloud account.

Value

The corresponding study Id.

[Package imotionsData version 1.8.5 Index]