getUploadSampleDataUrl {imotionsData}	R Documentation

Gets the endpoint url for a stimulus and a respondent (or segment)

Description

Gets the endpoint url for a stimulus and a respondent (or segment)

Usage

getUploadSampleDataUrl(connection, stimulus, endpoint)

Arguments

connection
As returned from connect function.

stimulus
A stimulus.

endpoint
either a respondent or a segment.

[Package imotionsData version 1.8.5 Index]