listRespondentSamples {imotionsData}	R Documentation

List samples available for a given respondent for a stimulus

Description

A sample is a record of data about a given respondent concerning a given stimulus. The data for a sample can get retrieved using imotionsData::getSampleData or imotionsData::getSampleDataFilePath.

Usage

listRespondentSamples(connection, studyId, stimulusId, respondentId)

Arguments

connection
As returned from connect function.

studyId
The study id

stimulusId
The stimulus for which you would like to list samples.

respondentId
The respondent to query

Details

Respondents for a given stimuli can be found using imotionsData::listStimulusRespondents

Value

A data frame containing the samples collected for a respondent from a stimulus within this study.

Examples

studies <- imotionsData::listStudies(connection)
studyId <- studies$id[1]
stimuli <- imotionsData::listStimuli(connection, studyId)
stimulusId <- stimuli$id[1]
respondents <- imotionsData::listStimulusRespondents(connection, studyId, stimulusId)
samples <- imotionsData::listRespondentSamples(connection, studyId, stimulusId, respondents$id[1])

[Package imotionsData version 1.8.5 Index]