listSampleRespondents {imotionsData}	R Documentation

List the samples and corresponding respondent id's for a stimulus

Description

This tells you which respondents have data for each sample in a given stimulus.

Usage

listSampleRespondents(connection, studyId, stimulusId)

Arguments

connection
As returned from connect function.

studyId
The study id

stimulusId
The stimulus for which you would like to list samples.

Details

A sample is a collection of data from a given respondent during exposure to a stimulus.

A sample source is the source of the sample data. It can be determined from the eventSourceType which may include Sensor, CustomEvent, or InternalEvent.

CustomEvent comes from a custom data import.

InternalEvent comes from internal events, such as mouse clicks and so on.

Samples for an individual respondent can be found using imotionsData::listRespondentSamples

Studies can be found with imotionsData::listStudies.

Value

A data frame containing a list of sample for the stimulus and the associated respondent id's.

Examples

studies <- imotionsData::listStudies(connection)
studyId <- studies$id[1]
stimuli <- imotionsData::listStimuli(connection, studyId)
sampleRespondents <- imotionsData::listSampleRespondents(connection, studyId, stimuli$id[1])

[Package imotionsData version 1.8.5 Index]