listSegmentSamples {imotionsData}	R Documentation

List samples available for a given segment for a stimulus

Description

A sample is a record of data about a given segment concerning a given stimulus. The data for a sample can get retrieved using imotionsData::getSampleData or imotionsData::getSampleDataFilePath.

Usage

listSegmentSamples(connection, studyId, stimulusId, segmentId)

Arguments

connection
As returned from connect function.

studyId
The study id

stimulusId
The stimulus for which you would like to list samples.

segmentId
The segment to query

Details

Segments for a given study can be found using imotionsData::listSegments

Value

A data frame containing the samples collected for a segment from a stimulus within this study.

Examples

studies <- imotionsData::listStudies(connection)
studyId <- studies$id[1]
stimuli <- imotionsData::listStimuli(connection, studyId)
stimulusId <- stimuli$id[1]
segments <- imotionsData::listSegments(connection, studyId)
samples <- imotionsData::listSegmentSamples(connection, studyId, stimulusId, segments$id[1])

[Package imotionsData version 1.8.5 Index]