listStimuli {imotionsData}	R Documentation

List stimuli within study

Description

More detailed information about a stimulus can be retrieved using imotionsData::getStimulus.

Usage

listStimuli(connection, studyId)

Arguments

connection
As returned from connect function.

studyId
The study id

Details

Studies can be found with imotionsData::listStudies.

Value

A data frame containing the names and ids of stimuli within this study.

Examples

studies <- imotionsData::listStudies(connection)
stimuli <- imotionsData::listStimuli(connection, studies$id[1])

[Package imotionsData version 1.8.5 Index]