listStimulusRespondents {imotionsData}	R Documentation

List the respondents that were shown a specific stimulus.

Description

List the respondents that were shown a specific stimulus.

Usage

listStimulusRespondents(connection, studyId, stimulusId)

Arguments

connection
As returned from connect function.

studyId
The study id

stimulusId
The stimulus for which you would like to list respondents

Value

A data frame containing respondents exposed to the stimulus

Examples

studies <- imotionsData::listStudies(connection)
stimuli <- imotionsData::listStimuli(connection, studies$id[1])
respondents <- imotionsData::listStimulusRespondents(connection, studies$id[1], stimuli$name[1])

[Package imotionsData version 1.8.5 Index]