privateUploadSampleDataFile {imotionsData}	R Documentation

Uploads sample data for a given stimulus and a respondent (or segment).

Description

Uploads sample data for a given stimulus and a respondent (or segment).

Usage

privateUploadSampleDataFile(
  connection,
  dataFileName,
  signalNames,
  stimulus,
  endpoint,
  sampleName,
  isMetric,
  overwrite
)

Arguments

connection
As returned from connect function

dataFileName
The data file containing the signals or metrics to upload

signalNames
Names of the signals or metrics

stimulus
A stimulus

endpoint
A respondent or a segment

sampleName
The name for the new sample

isMetric
logical indicating that the data being uploaded is a metric

overwrite
Overwrite if a sample with the same name exists.

[Package imotionsData version 1.8.5 Index]