stopSubjectNotApplicable {imotionsData}	R Documentation

Indicate that R markdown knitting should be stopped since the current stimulus/respondent/segment is not applicable to this notebook.

Description

Indicate that R markdown knitting should be stopped since the current stimulus/respondent/segment is not applicable to this notebook.

Usage

stopSubjectNotApplicable()

[Package imotionsData version 1.8.5 Index]