unloadStudies {imotionsData}	R Documentation

Unloads the studies that have been loaded in the current session

Description

Unloads the studies that have been loaded in the current session

Usage

unloadStudies()

Value

TRUE if success

[Package imotionsData version 1.8.5 Index]