uploadSampleData {imotionsData}	R Documentation

Creates a new Sample for a stimuli and respondent (or segment) pair in a study. Can upload both signals (with a TimeStamp column) or metrics (single values, with no TimeStamp)

Description

After processing, the sample can then be viewed in the Cloud, or downloaded again through the R data library.

Usage

uploadSampleData(
  connection,
  studyId,
  stimulusId,
  sampleName,
  respondentId = NULL,
  segmentId = NULL,
  data = NULL,
  file = NULL,
  isMetric = FALSE,
  overwrite = FALSE
)

Arguments

connection
As returned from connect function.

studyId
The study in question

stimulusId
The stimulus of interest

sampleName
The name of the sample that you would like to create

respondentId
The respondent that the sample is for (only if segmentId = NULL)

segmentId
The segment that the sample is for (only if respondentId = NULL)

data
A data.frame containing the sample data to upload.

file
The path to a gzipped, tab-delimited file containing the sample data to upload.

isMetric
logical indicating if the uploaded sample is a signal (with a TimeStamp column) or a metric (single value, without TimeStamp)

overwrite
logical indicating that the sample can be overwritten, if already present otherwise uploadSampleData will result in an error.

Details

Samples are a portion of the data gathered for a given respondent during exposure to a stimuli, from a given sample source.

Value

The Sample that was created

[Package imotionsData version 1.8.5 Index]